airflow_dags/requirements.txt
apache-airflow==2.8.1
boto3
pendulum